basic componentisation
