+++ frontend/src/components/LoggedIn/Meetings/OrganisationMeetingsList.jsx
@@ -1,7 +1,60 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import styles from './OrganisationMeetingsList.module.css';
 
-const OrganisationMeetingsList = ({ meetings, goToMeeting }) => {
+const DUMMY_DATA = [
+  {
+    id: 0,
+    title: 'meeting 1',
+    date: '10-06-2024'
+  },
+  {
+    id: 1,
+    title: 'meeting 2',
+    date: '10-06-2024'
+  },
+  {
+    id: 3,
+    title: 'meeting 3',
+    date: '10-06-2024'
+  }
+];
+
+const OrganisationMeetingsList = ({ organisationName, goToMeeting }) => {
+  const [meetings, setMeetings] = useState([]);
+
+  const getOrganisationMeetings = async () => {
+    try {
+      const response = await fetch(`${process.env.REACT_APP_BACKEND_URL}/get-organisation-meetings`, {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json'
+        },
+        body: JSON.stringify({
+          name: organisationName
+        }),
+        credentials: 'include'
+      });
+
+      if (!response.ok) {
+        const errorResponse = await response.json();
+        const errorText = 'An error occurred fetching your organisations.';
+        throw new Error(errorText);
+      }
+
+      const data = await response.json();
+      setMeetings(data.organisations);
+    } catch (error) {
+      console.log('ERROR');
+    }
+
+    // to be removed after endpoint works
+    setMeetings(DUMMY_DATA);
+  };
+
+  useEffect(() => {
+    getOrganisationMeetings();
+  }, [organisationName]);
+
   return (
     <div>
       {meetings.length === 0 ? (
