+++ frontend/src/components/LoggedIn/Meetings/TeamMeetingsList.jsx
@@ -1,7 +1,65 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import styles from './TeamMeetingsList.module.css';
 
-const TeamMeetingsList = ({ meetings, goToMeeting }) => {
+const DUMMY_DATA = [
+  {
+    id: 0,
+    title: 'meeting 1',
+    date: '10-06-2024'
+  },
+  {
+    id: 1,
+    title: 'meeting 2',
+    date: '10-06-2024'
+  },
+  {
+    id: 3,
+    title: 'meeting 3',
+    date: '10-06-2024'
+  }
+];
+
+const TeamMeetingsList = ({ teamName, organisationName, goToMeeting }) => {
+  const [meetings, setMeetings] = useState([]);
+
+  const getTeamMeetings = async () => {
+    try {
+      const response = await fetch(`${process.env.REACT_APP_BACKEND_URL}/get-team-meetings`, {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json'
+        },
+        body: JSON.stringify({
+          name: teamName,
+          organisation: organisationName
+        }),
+        credentials: 'include'
+      });
+
+      if (!response.ok) {
+        const errorResponse = await response.json();
+        const errorText = 'An error occurred fetching your organisations.';
+        throw new Error(errorText);
+      }
+
+      const data = await response.json();
+      setMeetings(data.meetings);
+    } catch (error) {
+      console.log('ERROR');
+    }
+
+    // to be removed after endpoint works
+    setMeetings(DUMMY_DATA);
+  };
+
+  useEffect(() => {
+    if (teamName) {
+      getTeamMeetings();
+    } else {
+        setMeetings([])
+    }
+  }, [teamName]);
+
   return (
     <div>
       {meetings.length === 0 ? (
